Fix crash downloading large files due to the DownloadItemView
When downloading a large file the DownloadItemView is supposed to show
certain information about the item being downloaded, such as the name
and the current status. Additionally, it's also meant to show the URL
but, due to changes done recently in PR 9521 [1] (see issue in [2]),
Brave contains changes to only show the URL when hovering over the
download item view, to save space.

Turns out that these changes were causing a crash after a the change
below landed upstream and the reason seems to be that our chromium_src
override was wrongly redefining the Button symbol way beyond the scope
of download_item_view.cc, causing the crash.

This change includes the necessary headers from download_item_view.cc
to avoid this Button redefinition from propagating any further, and
also makes the most of the effort to fix some small things in the
override file, such as calling the parent methods of the overriden
methods from the Button class, or adding some extra documentation.

Last, this change also fixes 2 crashing browser tests:

  * BraveCheckClientDownloadRequestBaseBrowserTest.FilterRequest
  * SavePackageFinishedObserverBrowserTest.Success

[1] https://github.com/brave/brave-core/pull/9521
[2] https://github.com/brave/brave-browser/issues/1638

Chromium change:

https://source.chromium.org/chromium/chromium/src/+/03f006384e02fb387c177af07a3df9792e364e18

commit 03f006384e02fb387c177af07a3df9792e364e18
Author: Peter Kasting <pkasting@chromium.org>
Date:   Thu Jul 22 00:45:59 2021 +0000

    Clicking the download item context button should close any open menu.

    This is what other buttons in Chrome do: repeatedly clicking the button
    toggles it between open and closed, instead of flickering it closed and
    instantly reopening (the current behavior).

    Bug: 1031682